Beacon lookups are wrapped in a background task
By wrapping the beacon lookups in a background task the system allows
for a longer time to finish: a normal wakeup from a location event
only allows 10s of execution, within a background task we get 180s.

This is necessary so that the lookups also work when using a shaky
network connection on a slow device.