Moved beacon lookup to beacon handler
This way, the beacon scanner is operating only on the abstraction
level of CLBeacons and its only responsibility is to detect beacons.
The beacon handler on the other side is responsible for everything
that happens when a beacon was detected, namely looking it up and
notifiying the user/app about it.